+++ src/session_impl.hpp
@@ -207,7 +207,7 @@ namespace sdk {
         virtual amount get_min_fee_rate() const = 0;
         virtual amount get_default_fee_rate() const = 0;
         virtual uint32_t get_block_height() const = 0;
-        virtual amount get_dust_threshold() const = 0;
+        amount get_dust_threshold(const std::string& asset_id) const;
         virtual nlohmann::json get_spending_limits() const = 0;
         virtual bool is_spending_limits_decrease(const nlohmann::json& limit_details) = 0;