+++ src/transaction_utils.hpp
@@ -93,8 +93,8 @@ namespace sdk {
         const network_parameters& net_params, nlohmann::json& result, nlohmann::json& addressee);
 
     // Add an output from a JSON addressee
-    amount add_tx_addressee(session_impl& session, const network_parameters& net_params, nlohmann::json& result,
-        wally_tx_ptr& tx, nlohmann::json& addressee);
+    amount add_tx_addressee(session_impl& session, nlohmann::json& result, wally_tx_ptr& tx, nlohmann::json& addressee,
+        const std::string& asset_id_hex);
 
     vbf_t generate_final_vbf(byte_span_t input_abfs, byte_span_t input_vbfs, uint64_span_t input_values,
         const std::vector<abf_t>& output_abfs, const std::vector<vbf_t>& output_vbfs, uint32_t num_inputs);